requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=1.10.0
langchain>=0.1.0
langgraph>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
alembic>=1.12.0
redis>=5.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
aiofiles>=23.2.1
jinja2>=3.1.2
email-validator>=2.1.0
psutil>=5.9.0